+++ app/src/main/java/com/paw/key/domain/entity/walk/WalkStartEntity.kt
@@ -2,5 +2,5 @@ package com.paw.key.domain.entity.walk
 
 data class WalkStartEntity(
     val routeId: String,
-    val issuedAt: Int
+    val issuedAt: Long
 )
